Refactor out some copy and pasted code
