+++ frontend/app/controllers/Contributor.scala
@@ -27,7 +27,7 @@ import views.support.{PageInfo, ThankYouMonthlySummary}
 import scala.concurrent.Future
 import scala.util.Failure
 
-object Contributor extends Controller with ActivityTracking
+object Contributor extends Controller with ActivityTracking with PaymentGatewayErrorHandler
   with LazyLogging
   with CatalogProvider
   with StripeServiceProvider
@@ -116,31 +116,10 @@ object Contributor extends Controller with ActivityTracking
 
     val prpId = request.contributor.productRatePlanId
 
-      Ok(views.html.joiner.thankyouMonthly(
-        request.contributor,
-        ThankYouMonthlySummary(LocalDate.now(), request.contributor.charges.price.prices.head),
-        resolution
-      ))
+    Ok(views.html.joiner.thankyouMonthly(
+      request.contributor,
+      ThankYouMonthlySummary(LocalDate.now(), request.contributor.charges.price.prices.head),
+      resolution
+    ))
   }
-
-
-  //TODO: copy pasted from Joiner - refactor
-  private def handlePaymentGatewayError(e: PaymentGatewayError, userId: String, tier: String, tracking: List[(String, String)], country: String = "") = {
-
-    def handleError(code: String) = {
-      logger.warn(s"User $userId could not become $tier member due to payment gateway failed transaction: \n\terror=$e \n\tuser=$userId \n\ttracking=$tracking \n\tcountry=$country")
-      Forbidden(Json.obj("type" -> "PaymentGatewayError", "code" -> code))
-    }
-
-    e.errType match {
-      case Fraudulent => handleError("Fraudulent")
-      case TransactionNotAllowed => handleError("TransactionNotAllowed")
-      case DoNotHonor => handleError("DoNotHonor")
-      case InsufficientFunds => handleError("InsufficientFunds")
-      case RevocationOfAuthorization => handleError("RevocationOfAuthorization")
-      case GenericDecline => handleError("GenericDecline")
-      case _ => handleError("UknownPaymentError")
-    }
-  }
-
 }
